requirements.txt
absl-py==2.1.0
accelerate==1.2.0
addict==2.4.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-cors==0.7.0
aiosignal==1.3.1
airportsdata==20241001
alabaster==1.0.0
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
arrow==1.3.0
arxiv==2.1.3
astor==0.8.1
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.2.0
babel==2.16.0
backcall==0.2.0
beautifulsoup4==4.9.0
bibtexparser==1.4.2
blake3==1.0.2
bleach==6.2.0
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.0
cbor==1.0.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
class-registry==2.1.2
click==8.1.7
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
colorful==0.5.6
comm
compressed-tensors==0.8.1
contourpy==1.3.1
cycler==0.12.1
dashscope==1.22.1
dataclasses-json==0.5.14
datasets==2.19.0
debugpy
decorator
defusedxml==0.7.1
Deprecated==1.2.14
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
docopt==0.6.2
docutils==0.21.2
docx==0.2.4
duckduckgo_search==6.3.4
einops==0.8.0
et_xmlfile==2.0.0
exceptiongroup
executing
faiss-cpu==1.9.0
fake-useragent==1.5.1
fastapi==0.115.5
fastjsonschema==2.21.1
fastprogress==1.0.3
feedparser==6.0.11
filelock==3.16.1
FlagEmbedding
Flask==3.1.0
fonttools==4.54.1
free_proxy==1.1.3
frozenlist==1.5.0
fsspec==2024.3.1
gensim==4.3.3
gguf==0.10.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.24.0
google-api-python-client==2.158.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google_search_results==2.4.2
googleapis-common-protos==1.66.0
greenlet==3.1.1
gritql==0.1.5
grpcio==1.70.0
h11==0.14.0
h5py==3.12.1
html2text==2024.2.26
httpcore==1.0.6
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.5
humanfriendly==10.0
idna==3.10
ijson==3.3.0
imagesize==1.4.1
impact-factor==1.1.1
importlib_metadata
importlib_resources==6.4.5
iniconfig==2.0.0
inscriptis==2.5.0
interegular==0.3.3
ipykernel
ipython==8.12.3
ir_datasets==0.5.9
itsdangerous==2.2.0
jedi
Jinja2==3.1.4
jiter==0.7.1
joblib==1.4.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client
jupyter_core
jupyterlab_pygments==0.3.0
kiwisolver==1.4.7
langchain==0.3.7
langchain-cli==0.0.31
langchain-community==0.3.7
langchain-core==0.3.17
langchain-openai==0.2.8
langchain-text-splitters==0.3.2
langserve==0.3.0
langsmith==0.1.142
lark==1.2.2
llvmlite==0.44.0
lm-format-enforcer==0.10.9
lmdb==1.6.2
loguru==0.7.3
lxml==5.3.0
lz4==4.3.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.9.2
matplotlib-inline
matplotlib-venn==1.1.1
mdurl==0.1.2
mistral_common==1.5.2
mistune==3.1.2
molbloom==2.2.1
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
narwhals==1.13.4
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest_asyncio
networkx==3.4.2
nltk==3.9.1
numba==0.61.0
numexpr==2.10.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.54.4
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
orjson==3.10.11
outcome==1.3.0.post0
outlines==0.1.11
outlines_core==0.1.26
packaging
pandas==1.5.3
pandocfilters==1.5.1
paper-qa==1.1.1
paperscraper==0.2.14
parso
partial-json-parser==0.2.1.1.post5
peft==0.14.0
pexpect
pickleshare
pillow==10.4.0
pipreqs==0.5.0
platformdirs
pluggy==1.5.0
prettytable==3.12.0
primp==0.7.0
prometheus-fastapi-instrumentator==7.0.2
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
psutil==6.1.0
ptyprocess
pure_eval
py-cpuinfo==9.0.0
py-spy==0.4.0
pyarrow==18.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11==2.13.6
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pylatexenc==2.10
pymed_paperscraper==0.0.1
pyparsing==3.2.0
pypdf==5.1.0
PyPDF2==3.0.1
pyreadline3==3.5.4
PySocks==1.7.1
pystow==0.5.6
pytest==8.3.3
python-dateutil
python-dotenv==1.0.1
python-pptx==1.0.2
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.10.1
ray==2.41.0
rdkit==2024.3.6
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rmrkl==0.0.3
rouge_score==0.1.2
rpds-py==0.21.0
rsa==4.9
RXN4Chemistry==1.14.0
safetensors==0.4.5
scholarly==1.7.11
scikit-learn==1.6.0
scipy==1.13.1
seaborn==0.13.2
selenium==4.26.1
sentence-transformers==3.3.1
sentencepiece==0.2.0
sgmllib3k==1.0.0
shellingham==1.5.4
shortuuid==1.0.13
simple-loggers==1.0.5
six==1.16.0
skunk==1.3.0
smart-open==7.0.5
SmilesPE==0.0.3
smmap==5.0.1
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.6
Sphinx==8.1.3
sphinx-rtd-theme==3.0.1
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sql-manager==1.0.5
SQLAlchemy==2.0.35
sse-starlette==1.8.2
stack_data
starlette==0.41.2
streamlit==1.40.1
sympy==1.13.1
synspace==0.3.0
tenacity==8.5.0
thefuzz==0.22.1
threadpoolctl==3.5.0
tiktoken==0.7.0
tinycss2==1.4.0
tokenizers==0.20.1
toml==0.10.2
tomli==2.1.0
tomlkit==0.12.5
torch==2.5.1
torch-geometric==2.6.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.67.0
traitlets
transformers==4.45.2
trec-car-tools==2.6
trio==0.27.0
trio-websocket==0.11.1
triton==3.1.0
typer==0.9.4
types-python-dateutil==2.9.0.20241003
typing-inspect==0.9.0
typing_extensions
tzdata==2024.2
unimol_tools==0.1.2.post2
unlzw3==0.2.3
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.23.2
uvloop==0.21.0
virtualenv==20.29.1
vllm==0.6.6.post1
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
watchdog==6.0.0
watchfiles==1.0.4
wcwidth
webencodings==0.5.1
webrequests==1.0.8
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
wikipedia==1.4.0
wrapt==1.16.0
wsproto==1.2.0
xformers==0.0.28.post3
xgrammar==0.1.10
XlsxWriter==3.2.1
xxhash==3.5.0
yarg==0.1.9
yarl==1.17.1
zipp
zlib-state==0.1.9
